+++ src/js/index.js
@@ -1,4 +1,5 @@
 import { initLoginModal } from "../js/events/initLoginModal.js";
+import { initShippingModal } from "../js/events/initShippingModal.js";
 console.log(screen.width);
 console.log(document.documentElement.scrollWidth);
 
@@ -6,5 +7,6 @@ console.log(document.documentElement.scrollWidth);
 
 const init = () =>{
     initLoginModal()
+    initShippingModal()
 }
 init()
